Modification des hacheurs pour qu'il travaille sur 32bits

+++ src/main/java/com/nedjar/algo/hachage/Hacheur.java
@@ -3,5 +3,5 @@ package com.nedjar.algo.hachage;
 import com.nedjar.dataStructure.Tuple;
 
 public interface Hacheur {
-	public long hache(Tuple key);
+	public int hache(Tuple key);
 }
